requirements.txt
streamlit==1.35.0
python-dotenv==1.0.1
langchain==0.2.1
ollama==0.1.7
langchain-core==0.1.50
langchain-community==0.2.0
